for Sushil
Will need to do add the --enable-linux_multipage-folio option for
configure:
   ./regen.sh
   ./configure --enable-linux-multipage-folio ... what you were using before

this compiles cleanly now.

I've moved your code into it's own copy of the functions
 #if defined(LINUX_MULTIPAGE_FOLIO)
   your version of a function code
 #else
   original unmodified version of the function
 #endif

I commented out/#if 0 the code for "dynamic" detection.

In general, don't use "//" for comments -- it's not standard C.  Use
/* blah blah */

AFS_CHUNKSIZE is a predefined macro that takes an offset and returns the
correct size of the chunk depending on the current offset.  I think that
this could be changed.

You don't need to use the page->private to hold the order.  I think what
you want to do is to allocate the folio with the number (size) you want
to use and pass the folio to afs_linux_read_cache.  The folio structure
should have the size.  If you do this, create a different function that
is modeled off afs_linux_read_cache (afs_linux_read_cache_folio), that
takes a folio instead of just a single page.

Currently:

afs_linux_readahead - is passed a rac, which should have a number of pages to read ahead
  - has a loop where it calls readahead_page to return the "next page" from the rac.
    determines the offset
      - calls "get_dcache_readahead"
      - calls ==> afs_linux_read_cache passing the single page from the rac to fill that page

afs_linux_read_folio
  just takes the head page from the folio
  calls ==> afs_linux_readpage with just the single page
    calls ==> afs_linux_fillpage with just a single page
      calls ==> afs_linux_readpage_fastpath with just a single page
        calls ==> afs_linux_read_cache passing a single page

There is a new commit pending for master that changes afs_linux_readahead from using
readahead_page to readahead_folio. (which is available back to linux 5.17) -- this was needed to
get a clean compile on Linux 6.16 because 6.16 removed the readahead_page function. -- the pending
code still just uses single pages.

update your afs_linux_readahead to call readahead_folio instead of readahead_page:

    while(folio = readahead_folio(rac)) do {
        and instead of just grabbing the head page, use the full folio.
        -- see linux source code for fs/ext4/readpage.c function ext4_mpage_readpages
        call afs_linux_read_cache_folio (new function?) that fills the folio from the disk cache.

        the code will need to handle a folio instead of a page (e.g. folio_get/folio_put instead of
        get_page/put_page, etc.)
    }