linux: convert aops->writepage to writepages
As part of the transition from pages to folios, the Linux 6.16 commit:
'fs: Remove aops->writepage' (6b0dfabb35550)

removes the writepage member from the address space operations and now
requires the writepages member.

An earlier Linux commit converted the underlying support of writepages
to use folios. The Linux 6.3 commit:
'fs: convert writepage_t callback to pass a folio' (d585bdbeb79aa)

requires the callback function passed to write_cache_pages() to expect a
folio instead of a page.

Since the aops->writepages member predates Linux 2.6.18, testing for
its presence is not an accurate check. Instead, we check to see if the
write_cache_pages() function takes a callback that uses folios (the
d585bdbeb79aa commit).

To implement this, afs_linux_write_pages() is added for
afs_file_aops.writepages. This function calls write_cache_pages(), which
uses the new afs_linux_writefolio_cb() callback to process the folios.
The callback's implementation is based on the logic from
afs_linux_writepage() but is updated to operate directly on folios. This
ensures that multi-page folios are handled correctly, rather than only
processing the head page of a folio.

Note, this change is retroactive back to linux 6.3.

Note, According to the Linux documentation for write_cache_pages, the
return value from the writepages should either be 0, or a negative error
code.  However the existing writepage code can return a positive value
(the residual for a partial write). For this commit, the existing
behavior of returning a residual is preserved.

Change-Id: I3abfe024c0f37d3c112bb9581e93388905763ede